sp_weights

Struct FeePolynomial

pub struct FeePolynomial<Balance> { /* private fields */ }

A list of coefficients that represent a polynomial.

Can be evaluated at a specific u64 to get the fee. The evaluations happens by summing up all term results. The order of the coefficients matters since it uses saturating arithmetic. This struct does therefore not model a polynomial in the mathematical sense (polynomial ring).

For visualization purposes, the formulas of the unsigned terms look like:

(c[0].frac * x^(c[0].degree) + c[0].integer * x^(c[0].degree))
(c[1].frac * x^(c[1].degree) + c[1].integer * x^(c[1].degree))
...

Depending on the value of c[i].negative, each term is added or subtracted from the result. The result is initialized as zero.

Implementations

impl<Balance> FeePolynomial<Balance>
where Balance: BaseArithmetic + From<u32> + Copy + Unsigned,

pub fn eval(&self, x: u64) -> Balance

Evaluate the polynomial at a specific x.