Struct sp_weights::Weight

pub struct Weight { /* private fields */ }

Implementations

impl Weight

pub const fn set_ref_time(self, c: u64) -> Self

Set the reference time part of the weight.

pub const fn ref_time(&self) -> u64

Return the reference time part of the weight.

pub fn ref_time_mut(&mut self) -> &mut u64

Return a mutable reference time part of the weight.

pub const MAX: Self = _

pub fn min(&self, other: Self) -> Self

Get the conservative min of self and other weight.

pub fn max(&self, other: Self) -> Self

Get the aggressive max of self and other weight.

pub fn try_add(&self, other: &Self, limit: &Self) -> Option<Self>

Try to add some other weight while upholding the limit.

pub const fn from_ref_time(ref_time: u64) -> Self

Construct Weight with reference time weight.

pub const fn saturating_add(self, rhs: Self) -> Self

Saturating Weight addition. Computes self + rhs, saturating at the numeric bounds of all fields instead of overflowing.

pub const fn saturating_sub(self, rhs: Self) -> Self

Saturating Weight subtraction. Computes self - rhs, saturating at the numeric bounds of all fields instead of overflowing.

pub const fn saturating_mul(self, scalar: u64) -> Self

Saturating Weight scalar multiplication. Computes self.field * scalar for all fields, saturating at the numeric bounds of all fields instead of overflowing.

pub const fn saturating_div(self, scalar: u64) -> Self

Saturating Weight scalar division. Computes self.field / scalar for all fields, saturating at the numeric bounds of all fields instead of overflowing.

pub const fn saturating_pow(self, exp: u32) -> Self

Saturating Weight scalar exponentiation. Computes self.field.pow(exp) for all fields, saturating at the numeric bounds of all fields instead of overflowing.

pub fn saturating_accrue(&mut self, amount: Self)

Increment Weight by amount via saturating addition.

pub const fn checked_add(&self, rhs: &Self) -> Option<Self>

Checked Weight addition. Computes self + rhs, returning None if overflow occurred.

pub const fn checked_sub(&self, rhs: &Self) -> Option<Self>

Checked Weight subtraction. Computes self - rhs, returning None if overflow occurred.

pub const fn checked_mul(self, scalar: u64) -> Option<Self>

Checked Weight scalar multiplication. Computes self.field * scalar for each field, returning None if overflow occurred.

pub const fn checked_div(self, scalar: u64) -> Option<Self>

Checked Weight scalar division. Computes self.field / scalar for each field, returning None if overflow occurred.

pub const fn zero() -> Self

Return a Weight where all fields are zero.

pub const fn any_gt(self, other: Self) -> bool

Returns true if any of self’s constituent weights is strictly greater than that of the other’s, otherwise returns false.

pub const fn all_gt(self, other: Self) -> bool

Returns true if all of self’s constituent weights is strictly greater than that of the other’s, otherwise returns false.

pub const fn any_lt(self, other: Self) -> bool

Returns true if any of self’s constituent weights is strictly less than that of the other’s, otherwise returns false.

pub const fn all_lt(self, other: Self) -> bool

Returns true if all of self’s constituent weights is strictly less than that of the other’s, otherwise returns false.

pub const fn any_gte(self, other: Self) -> bool

Returns true if any of self’s constituent weights is greater than or equal to that of the other’s, otherwise returns false.

pub const fn all_gte(self, other: Self) -> bool

Returns true if all of self’s constituent weights is greater than or equal to that of the other’s, otherwise returns false.

pub const fn any_lte(self, other: Self) -> bool

Returns true if any of self’s constituent weights is less than or equal to that of the other’s, otherwise returns false.

pub const fn all_lte(self, other: Self) -> bool

Returns true if all of self’s constituent weights is less than or equal to that of the other’s, otherwise returns false.

pub const fn any_eq(self, other: Self) -> bool

Returns true if any of self’s constituent weights is equal to that of the other’s, otherwise returns false.