Crate sp_weights

Primitives for transaction weighting.

Modules

constants

Structs

ConstantMultiplier

Implementor of WeightToFee that uses a constant multiplier.

FeePolynomial

A list of coefficients that represent a polynomial.

FixedFee

Implementor of WeightToFee such that it maps any unit of weight to a fixed fee.

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

RuntimeDbWeight

The weight of database operations that the runtime can invoke.

Weight

WeightMeter

Meters consumed weight and a hard limit for the maximal consumable weight.

WeightToFeeCoefficient

One coefficient and its position in the WeightToFee.

Traits

WeightToFee

A trait that describes the weight to fee calculation.

WeightToFeePolynomial

A trait that describes the weight to fee calculation as polynomial.

Type Aliases

NoFee

An implementation of WeightToFee that collects no fee.

WeightToFeeCoefficients

A list of coefficients that represent a polynomial.