Crate sp_weights

Primitives for transaction weighting.

Modules

constants

Structs

ConstantMultiplier

Implementor of WeightToFee that uses a constant multiplier.

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

OldWeight

The old weight type.

RuntimeDbWeight

The weight of database operations that the runtime can invoke.

Weight

WeightMeter

Meters consumed weight and a hard limit for the maximal consumable weight.

WeightToFeeCoefficient

One coefficient and its position in the WeightToFee.

Traits

WeightToFee

A trait that describes the weight to fee calculation.

WeightToFeePolynomial

A trait that describes the weight to fee calculation as polynomial.

Type Definitions

WeightToFeeCoefficients

A list of coefficients that represent one polynomial.