Crate sp_weights

Primitives for transaction weighting.

Latest machine specification used to benchmark are:

Digital Ocean: ubuntu-s-2vcpu-4gb-ams3-01

2x Intel(R) Xeon(R) CPU E5-2650 v4 @ 2.20GHz

4GB RAM

Ubuntu 19.10 (GNU/Linux 5.3.0-18-generic x86_64)

rustc 1.42.0 (b8cedc004 2020-03-09)

Modules

constants

Structs

ConstantMultiplier

Implementor of WeightToFee that uses a constant multiplier.

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

OldWeight

The old weight type.

RuntimeDbWeight

The weight of database operations that the runtime can invoke.

Weight

WeightMeter

Meters consumed weight and a hard limit for the maximal consumable weight.

WeightToFeeCoefficient

One coefficient and its position in the WeightToFee.

Traits

WeightToFee

A trait that describes the weight to fee calculation.

WeightToFeePolynomial

A trait that describes the weight to fee calculation as polynomial.

Type Definitions

WeightToFeeCoefficients

A list of coefficients that represent one polynomial.