Trait sp_weights::WeightToFeePolynomial

pub trait WeightToFeePolynomial {
    type Balance: BaseArithmetic + From<u32> + Copy + Unsigned;

    fn polynomial() -> WeightToFeeCoefficients<Self::Balance>;
}

A trait that describes the weight to fee calculation as polynomial.

An implementor should only implement the polynomial function.

Required Associated Types

type Balance: BaseArithmetic + From<u32> + Copy + Unsigned

The type that is returned as result from polynomial evaluation.

Required Methods

fn polynomial() -> WeightToFeeCoefficients<Self::Balance>

Returns a polynomial that describes the weight to fee conversion.

This is the only function that should be manually implemented. Please note that all calculation is done in the probably unsigned Balance type. This means that the order of coefficients is important as putting the negative coefficients first will most likely saturate the result to zero mid evaluation.